Reuse attribute structure array if safe to do so
The value of the NTNDArray attribute field is reused when this is safe.
The shared vector of structure arrays is reused when the use count is 1
via the reuse() function. The individual structures are reused when in
addition their use count is 1.

This has required a rewrite of the attribute code. Reference semantics
create function have been replace by value semantic set functions.

Signed-off-by: Dave Hickin <david.hickin@diamond.ac.uk>